Truely parse the remote CSV.
The previous version only pretended to parse the remote CSV,
always assuming the tram-line would be two characters in length.
That is probably just fine for trams, but for buses it fails.

Busses will be output from this URL:

* https://steve.fi/Helsinki/Tram-API/api.cgi?id=1160107
  * https://beta.reittiopas.fi/pysakit/HSL:1160107

They show:

        436K,18:52:11,Elielinaukio-Martinlaakso-Juvanmalmi-Kalajärvi
        63,18:53:56,Elielinaukio-Maunula-Paloheinä
        322,18:54:00,Elielinaukio-Vantaankoski
        42,18:56:19,Kamppi-Maununneva-Kannelmäki

As you can see the bus-lines have IDs such as "42", "322", "436K"
which are longer than two characters.

We now cope with that :)