Overhauled user-interface significantly.
This large commit updates the web-based user-interface, such
that it uses a series of bootstrap-tabs, rather than a list of
paragraphs to view/edit the options.

Despite the change the code has actually gotten smaller, due to
a lack of formatted-rows.

The `display-mode` setup has been simplified to use a series
of radio-buttons which is reasonably efficient too.

This closes #13.