Allow the web-UI to set the backlight schedule.
Rather than being hardwired to enable the backlight at 6AM,
and disable it at 10PM, we now let the user enable/disable
a schedule.

We have a checkbox to enable the schedule, and if enabled the
user can choose an on & off hour.  There is no sanity-checking
that the times are identical though!

This closes #16.